Refactor SAML Response spec helpers
**Why**: Combine module methods into a single
SamlResponseDoc class.
